rename queue provider file
